feat(elasticsearch-sync): support multiple sites side by side
To actually sync two different StackExchange sites side
by side with ninya.io we need to make sure to prefix
the id with the site name to be unique.

We also need to make the index name configureable because
we have to create a new index for such breaking changes.